remove unnecessary cast from double to int

+++ HA4/euler.c
@@ -55,7 +55,7 @@ SimulationSettings getHandle(SimulationCalculation simulationCalculation) {
 
 int calculateSimulation(void *data, const Plot *plot, StateToPlotWriter writeToPlot) {
     SimulationSettings *settings = data;
-    int timeStepCount = (int) ceil(settings->duration / settings->stepSize);
+    int timeStepCount = ceil(settings->duration / settings->stepSize);
 
     SimulationState currentState = settings->initialState;
